[refactor] #45 Floating Button 디자인 수정 및 기능 구현
